Lots of improvements and upgrade wgpu version

+++ rshader/src/dynamic_shaders.rs
@@ -48,9 +48,9 @@ fn concat_file_contents<'a, I: Iterator<Item = &'a PathBuf>>(filenames: I) -> io
 }
 
 pub struct ShaderSet {
-    vertex: Option<Vec<u8>>,
-    fragment: Option<Vec<u8>>,
-    compute: Option<Vec<u8>>,
+    vertex: Option<Vec<u32>>,
+    fragment: Option<Vec<u32>>,
+    compute: Option<Vec<u32>>,
 
     input_attributes: Vec<wgpu::VertexAttributeDescriptor>,
     layout_descriptor: Vec<wgpu::BindGroupLayoutEntry>,
@@ -193,13 +193,13 @@ impl ShaderSet {
         &self.input_attributes[..]
     }
 
-    pub fn vertex(&self) -> &[u8] {
+    pub fn vertex(&self) -> &[u32] {
         self.vertex.as_ref().unwrap()
     }
-    pub fn fragment(&self) -> &[u8] {
+    pub fn fragment(&self) -> &[u32] {
         self.fragment.as_ref().unwrap()
     }
-    pub fn compute(&self) -> &[u8] {
+    pub fn compute(&self) -> &[u32] {
         self.compute.as_ref().unwrap()
     }
 }
